build(changesets): :package: update changelog.ts
