--- utils/changesets/CHANGELOG.md
@@ -1,35 +1,0 @@
-# @slardar/changesets
-
-## 0.0.2
-
-### Patch Changes
-
-- [455baad48d4131b115a443a1032b95ad36f8cf9a](https://github.com/cjinhuo/mono-sdk-boilerplate/commit/455baad48d4131b115a443a1032b95ad36f8cf9a) @cjinhuo@qq.com - 2025-07-11 => test changelog
-
-## 0.0.2
-
-### Patch Changes
-
-- [e79524df](https://code.byted.org/appmonitor/slardar_sdk/commit/e79524df8c7a8eb66b408a718d7e12a5cb123423) @chenjinhuo@bytedance.com - 2023-05-09 => 在 changeset_version 添加参数 --filter,可以定向更新某个单包,参数格式符合 glob,例子:--filter='@slardar/sdk-hybrid' 或者 --filter='@slardar/\*'
-
-## 0.0.2-alpha.0
-
-### Patch Changes
-
-- [e79524df](https://code.byted.org/appmonitor/slardar_sdk/commit/e79524df8c7a8eb66b408a718d7e12a5cb123423) @chenjinhuo@bytedance.com - 2023-05-09 => 在 changeset_version 添加参数 --filter,可以定向更新某个单包,参数格式符合 glob,例子:--filter='@slardar/sdk-hybrid' 或者 --filter='@slardar/\*'
-
-## 0.0.1
-
-### Patch Changes
-
-- [3fcf88fa](https://code.byted.org/appmonitor/slardar_sdk/commit/3fcf88fada8c8dbced25acbc0f64103c3f216bf1) @chenjinhuo@bytedance.com - 2023-04-18 => - add extra param 'message-by-summary' for commit message by summary automatically
-- add extra param 'git-add' to control whether 'git add .' or not
-- [384edfb7](https://code.byted.org/appmonitor/slardar_sdk/commit/384edfb732716c8cfed72ab0fbb9a22a999cdf3b) @pengli.coder@bytedance.com - 2023-04-18 => feat(changeset): add auto commit
-
-## 0.0.1-alpha.0
-
-### Patch Changes
-
-- [3fcf88fa](https://code.byted.org/appmonitor/slardar_sdk/commit/3fcf88fada8c8dbced25acbc0f64103c3f216bf1) @chenjinhuo@bytedance.com - 2023-04-18 => - add extra param 'message-by-summary' for commit message by summary automatically
-- add extra param 'git-add' to control whether 'git add .' or not
-- [384edfb7](https://code.byted.org/appmonitor/slardar_sdk/commit/384edfb732716c8cfed72ab0fbb9a22a999cdf3b) @pengli.coder@bytedance.com - 2023-04-18 => feat(changeset): add auto commit
