+++ utils/changesets/src/changelog.ts
@@ -1,55 +1,28 @@
 import type { ModCompWithPackage, NewChangesetWithCommit } from '@changesets/types'
-import dayjs from 'dayjs'
-import { execa } from 'execa'
+import { formatGitCommitId, getGitRemoteUrl, getInfoByCommitId } from './helper'
 
-async function getInfoByCommitId(commitId) {
-	const splitChar = '__'
-	const res = await execa('git', ['show', `--pretty=%ae${splitChar}%cd${splitChar}%H`, commitId], { stdio: 'pipe' })
-	const match = res.stdout.match(/^(.*?)\n/m)
-	if (!match) {
-		throw new Error(`commitId ${commitId} not found`)
+export async function getReleaseLine(newChangesetWithCommit: NewChangesetWithCommit) {
+	if (!newChangesetWithCommit.commit) {
+		throw new Error('CommitId Not Found From Changeset')
 	}
-	const [email, date, intactHash] = match[1].split(splitChar)
-	return { email, date: dayjs(date).format('YYYY-MM-DD'), intactHash }
-}
-
-async function getRemoteUrl() {
-	try {
-		const res = await execa('git', ['config', '--get', 'remote.origin.url'], { stdio: 'pipe' })
-		let url = res.stdout.trim()
-		// 转换SSH URL为HTTPS URL
-		if (url.startsWith('git@')) {
-			url = url.replace(/^git@([^:]+):/, 'https://$1/')
-		}
-		// 移除.git后缀
-		url = url.replace(/\.git$/, '')
-		return url
-	} catch (error) {
-		return 'https://github.com/unknown/repo' // 默认值
-	}
-}
-
-export async function getReleaseLine(changeset: NewChangesetWithCommit) {
-	const { email, intactHash } = await getInfoByCommitId(changeset.commit)
-	const date = dayjs().format('YYYY-MM-DD')
-	const remoteUrl = await getRemoteUrl()
+	const { email, date, intactHash } = await getInfoByCommitId(newChangesetWithCommit.commit)
+	const remoteUrl = await getGitRemoteUrl()
 	// - [c0b7864a4] @chenjinhuo - 2023.03.12:新增某某字段
-	return `- [${changeset.commit}](${remoteUrl}/commit/${intactHash}) @${email} - ${date} => ${changeset.summary}`
+	return `- [${formatGitCommitId(newChangesetWithCommit.commit)}](${remoteUrl}/commit/${intactHash}) @${email} - ${date} => ${newChangesetWithCommit.summary}`
 }
 
 export async function getDependencyReleaseLine(
-	_changesets: NewChangesetWithCommit[],
+	newChangesetWithCommits: NewChangesetWithCommit[],
 	_dependenciesUpdated: ModCompWithPackage[]
 ) {
-	// self changelog has generated by getReleaseLine
-	if (!_changesets.length) return ''
+	if (!newChangesetWithCommits.length) return ''
 
 	const releaseLines: string[] = []
-	for (const changeset of _changesets) {
+	for (const changeset of newChangesetWithCommits) {
 		const releaseLine = await getReleaseLine(changeset)
 		releaseLines.push(releaseLine)
 	}
 	const dependenciesUpdated = _dependenciesUpdated[0]
 	const updatedBy = `- Updated By ${dependenciesUpdated.name}: ${dependenciesUpdated.oldVersion}->${dependenciesUpdated.newVersion}`
-	return updatedBy + '\n' + releaseLines.map((v) => `  ${v}`).join('\n')
+	return `${updatedBy}\n${releaseLines.map((v) => `  ${v}`).join('\n')}`
 }
